[Refactor] 파일 오타사항, 헤더관리 효율화, useCredit 업데이트방식 수정
