+++ src/hooks/list/useCredit.jsx
@@ -1,16 +1,16 @@
 import { useEffect } from 'react';
 import useSWR, { mutate } from 'swr';
 
-// 로컬스토리지에서 크레딧 현황 가져오는 훅
+// 로컬스토리지에서 크레딧 현황 가져오는 함수
 const fetchCredit = () => {
     const credit = parseInt(localStorage.getItem('credit'), 10) || 0;
     return credit;
 };
 
-// SWR을 사용하여 크레딧이 변경됐을 때 1초 후 업데이트
+// SWR을 사용하여 크레딧이 변경된 후 실행되면 업데이트
 const useCredit = () => {
-    const { data: credit } = useSWR('credit', fetchCredit, { refreshInterval: 1000 });
-
+    const { data: credit } = useSWR('credit', fetchCredit);
+    
     useEffect(() => {
         const updateCredit = () => {
             const currentCredit = fetchCredit();
@@ -24,10 +24,10 @@ const useCredit = () => {
         window.addEventListener('storage', updateCredit);
 
         return () => {
-            window.removeEventListener('storage', updateCredit); 
+            window.removeEventListener('storage', updateCredit);
         };
-    }, []);
-
+    }, [credit]);
+  
     return credit;
 };
