+++ src/pages/list/components/DonationList.jsx
@@ -4,7 +4,7 @@ import { getDonations } from '../../../api/donations';
 import DonationItem from './DonationItem';
 import rightBtnIcon from '../../../assets/icon/btn_pagination_right.svg';
 import lefgBtnIcon from '../../../assets/icon/btn_pagination_left.svg';
-import useScrollTo from '../../../hooks/useScrollTo ';
+import useScrollTo from '../../../hooks/useScrollTo';
 import usePagination from '../../../hooks/usePagination';
 import RefreshButton from '../../../components/RefreshButton';
