+++ src/pages/my/MyPage.jsx
@@ -1,6 +1,5 @@
 import React, { useEffect, useState } from 'react';
 import styled from 'styled-components';
-import Header from '../../components/Header';
 import InterestedIdols from './components/InterestedIdols';
 import AddInterestedIdols from './components/AddInterestedIdols';
 import { getIdols } from '../../api/idols';
@@ -96,7 +95,6 @@ const MyPage = () => {
     return (
         <ErrorBoundary FallbackComponent={ErrorPage}>
             <StyledMyPage>
-                <Header />
                 <MyStateContext.Provider value={{ datas, selectedDatas, checkedIdols }}>
                     <MyDispatchContext.Provider value={{ setDatas, setSelectedDatas, setCheckedIdols }}>
                         <InterestedIdols />
